APIs/jellyfin.local/v1/python-aiohttp-client/openapi_server/controllers/system_controller.py
from typing import List, Dict
from aiohttp import web

from openapi_server.models.end_point_info import EndPointInfo
from openapi_server.models.log_file import LogFile
from openapi_server.models.public_system_info import PublicSystemInfo
from openapi_server.models.system_info import SystemInfo
from openapi_server.models.wake_on_lan_info import WakeOnLanInfo
from openapi_server import util


async def get_endpoint_info(request: web.Request, ) -> web.Response:
    """Gets information about the request endpoint.

    


    """
    return web.Response(status=200)


async def get_log_file(request: web.Request, name) -> web.Response:
    """Gets a log file.

    

    :param name: The name of the log file to get.
    :type name: str

    """
    return web.Response(status=200)


async def get_ping_system(request: web.Request, ) -> web.Response:
    """Pings the system.

    


    """
    return web.Response(status=200)


async def get_public_system_info(request: web.Request, ) -> web.Response:
    """Gets public information about the server.

    


    """
    return web.Response(status=200)


async def get_server_logs(request: web.Request, ) -> web.Response:
    """Gets a list of available server log files.

    


    """
    return web.Response(status=200)


async def get_system_info(request: web.Request, ) -> web.Response:
    """Gets information about the server.

    


    """
    return web.Response(status=200)


async def get_wake_on_lan_info(request: web.Request, ) -> web.Response:
    """Gets wake on lan information.

    


    """
    return web.Response(status=200)


async def post_ping_system(request: web.Request, ) -> web.Response:
    """Pings the system.

    


    """
    return web.Response(status=200)


async def restart_application(request: web.Request, ) -> web.Response:
    """Restarts the application.

    


    """
    return web.Response(status=200)


async def shutdown_application(request: web.Request, ) -> web.Response:
    """Shuts down the application.

    


    """
    return web.Response(status=200)
